Eurostat migration data

These data sets have been downloaded from: the publicly available Eurostat data base on population: https://ec.europa.eu/eurostat/web/main/data/database, data bases 'migr' and 'demo'.

These are migration statistics for selected EU countries, and population sizes. The data relate to 2010-2019 period.

Shiny app made by Maciej Jan Dańko: https://maciej-jan-danko.shinyapps.io/HMigD_Shiny_App_I/, These data contain the assumed accuracy of the data sources.

Shiny app made by Maciej Jan Dańko: https://maciej-jan-danko.shinyapps.io/undercountmigscores/ These data contain undercounting scores for emigration and immigration based on 30% metadata and 70% model scores.

Data on distance between countries from the CEPII database cepii.fr/CEPII/.

Usage

mig_data

Format

Each is a tibble with two variables and three observations

Details

mig_data are a data frame/tibble.

Examples

mig_data
#> # A tibble: 7,020 × 24
#>    SEX    Year Sending  Receiving Sendin…¹ Recei…² Sendi…³ Recei…⁴ Corri…⁵  Emig
#>    <chr> <dbl> <chr>    <chr>     <chr>    <chr>   <chr>   <chr>   <chr>   <dbl>
#>  1 Total  2010 Bulgaria Belgium   BG       BE      BGR     BEL     BGR_BEL    NA
#>  2 Total  2010 Czechia  Belgium   CZ       BE      CZE     BEL     CZE_BEL    NA
#>  3 Total  2010 Denmark  Belgium   DK       BE      DNK     BEL     DNK_BEL   307
#>  4 Total  2010 Germany  Belgium   DE       BE      DEU     BEL     DEU_BEL    NA
#>  5 Total  2010 Estonia  Belgium   EE       BE      EST     BEL     EST_BEL    34
#>  6 Total  2010 Ireland  Belgium   IE       BE      IRL     BEL     IRL_BEL   431
#>  7 Total  2010 Greece   Belgium   GR       BE      GRC     BEL     GRC_BEL    NA
#>  8 Total  2010 Spain    Belgium   ES       BE      ESP     BEL     ESP_BEL  5835
#>  9 Total  2010 France   Belgium   FR       BE      FRA     BEL     FRA_BEL    NA
#> 10 Total  2010 Croatia  Belgium   HR       BE      HRV     BEL     HRV_BEL    NA
#> # … with 7,010 more rows, 14 more variables: Immi <dbl>, UI <fct>, AI <fct>,
#> #   UE <fct>, AE <fct>, Pop.Sending <dbl>, Pop.Receiving <dbl>, contig <dbl>,
#> #   comlang_off <dbl>, comlang_ethno <dbl>, dist <dbl>, distcap <dbl>,
#> #   distw <dbl>, distwces <dbl>, and abbreviated variable names ¹​Sending_iso2,
#> #   ²​Receiving_iso2, ³​Sending_iso3, ⁴​Receiving_iso3, ⁵​Corridor