Function to estimate migration model

The model estimates 'true', unobserved bilateral migration flows \(y_{ijt}\) from sending country \(i\) to receiving country \(j\) in year \(t\). It relies on data reported by receiving countries (immigration, \(z_{ijt}^R\)) and sending countries (emigration, \(z_{ijt}^S\)). It is based on methodology developed by Raymer et al. (2013). The data are Poisson-distributed with mean \(\mu_{ijt}^k\):

\(z_{ijt}^k\sim\mathrm{Poisson}\left(\mu_{ijt}^k\right)\).

The true migration flows then result from correcting the expected values of the data for biases represented by parameters \(\lambda\) and different accuracies of the sending and receiving countries' data \(\sigma\):

\(\log\mu_{ijt}^k \sim \mathrm{normal}\left(\log y_{ijt}+\lambda_{f(i,j)}, \sigma_{g(i,j)}^k\right)\).

The true flows are then modelled by using a mixed effects model that contains an autoregressive part with parameter \(\phi\) and a gravity-type model with populations of sending (\(X_{it}\)) and receiving (\(X_{jt}\)) countries as well as distance \(D_{ij}\) between them:

\(\log y_{ijt}^k \sim \mathrm{normal}\left(\psi_0 + \psi_{1,ij}\log y_{ijt-1} + \beta_1 \log X_{it} + \beta_2 \log X_{jt} + \beta_3 D_{ij}, \sigma_y\right)\)

Note, that this specification requires that there are no fully unobserved flows in the model.

Usage

migmod_m(
  df,
  sending = c("SE", "FI", "IT", "PL"),
  receiving = c("SE", "FI", "IT", "PL"),
  years = 2010:2019,
  ref.country = "SE",
  ...
)

Arguments

df

A data frame with migration data for sending countries, receiving countries, data quality measures and covariates

sending

A vector of ISO2 codes with sending countries to be used in the model (recommended to keep SE and FI)

receiving

A vector of ISO2 codes with receiving countries to be used in the model (recommended to keep SE and FI)

years

A vector of years to be used in the model (2010 to 2019).

ref.country

A character with an ISO2 code receiving country which immigration data are is used as a reference. This assumes that the data reported by this country are not biased. This assumption ensures identification of model parameters. It is strongly recommended to use Sweden ("SE") as a reference.

...

Arguments passed to function sampling from package rstan (e.g. iter, chains) or data_2_standata.

Value

An object of class stanfit returned by rstan::sampling

References

Raymer J., Wiśniowski A., Forster J. J., Smith P. W. & Bijak J. (2013). Integrated modeling of European migration. Journal of the American Statistical Association, 108(503), 801-819. https://doi.org/10.1080/01621459.2013.789435.

See also